retrieving the metadata from OCI while runtime
introducing `get_metadata` and `run_describe` which allows to get all metadata from a "description run" to be present and queried.
the list for the type conversions is not yet complete, nor even close. additionally we might need to distinguish between precision and scale in order to have precise information about the target data type