Port diesel-oci to use the oracle crate internally
This removes our own oci-sys crate in favour of the oracle crate, that
already provides a high-level rust oracle database interface. All
existing tests are passing, but there are hand full of things that are
not perfect yet:

* We don't do any prepared statement caching yet due to
https://github.com/kubo/rust-oracle/issues/28
* The handling of insert statments with returning clauses is rather
hacky. It works fine, but it does probably more than required

On the other hand this removes a lot of highly unsafe low-level code +
simplifies quite a lot of the implementation (See the diff-stats for example)